qml: track the sidebar size in dp rather than application width ratio
Having the pannel size keeping the same aspect ratio get anoying when the user
makes big resizes. A user setting the size of a pannel to fit a given text would
see its setting broken when resizing down.

Size is tracked in "dp" to keep the sizes coherent if scale factor changes or
user resize the interface