contrib: gettext: Fix building for older macOS versions
Work around the fact that gettext/gnulib tests for functions
without using headers, by manually marking a bunch of functions
as unavailable.

The function openat is extra tricky, because the gnulib functions
themselves unconditionally use openat (expecting to either get it
from the system, or from a gnulib polyfill). When using that,
it would see both declarations of openat, both from the system headers
and from gnulib. In that case the system headers still indicated that
the function is unavailable on 10.7.

Rename the function while including the system header, so that we
only see the gnulib replacement declaration.