qt: use DialogsProvider::getInstance(m_intf) for exit
On KDE Plasma, the QuitSpy is triggered without the DialogsProvider
being initialized, which triggers the assertion in DialogsProvider
getInstance() method.

	vlc: ../../../../modules/gui/qt/dialogs/dialogs_provider.hpp:101:
	static DialogsProvider* DialogsProvider::getInstance(): Assertion `instance' failed.

We can provide the p_intf ourselves at initialization, but we need to
make sure it's correctly destroyed afterwards. Since it's destroyed
afterwards in ThreadCleanup, this will be correctly deallocated.

Fixes #29205