player: timer: handle live state
If live is true, then the length represents the seek range and the
position variable is the position within that seek range.

Don't interpolate position with length in that case.

The timer live variable could be used by UI to change sliders design
accordingly.

This will allow fixing the 3.0 regression where it was impossible to
seek within a seek range with live streams.