vout_subpictures: position regions in the video regardless of absolute flag
This was originally introduced in 95fd2226a438f1eb31e00fbc7e18d28822ce351f.
But any region that was positioned within the video dimensions need to be
positioned back in the display.
On the other hand an absolute region doesn't need to be positioned back if it
was not positioned within the video.