CI: do not build macOS out of tree
Since a72e41eef352af3c5e91589ab1f1515d03c5d168 the contribs are built
in the build tree, source in the source tree.
The CI jobs expects the generated tarball+checksum to be in the source tree.
So we build in the source tree as the other targets.

Fixes #29214