qml: try to not exceed `min(width, height) / 2` as radius in `ScrollBarExt.qml`
This does not effectively change anything, as the `Rectangle` clamp it
anyway, but it still worth to have. The style wants maximum radius in
this case.

I specifically did not use `Math.min(width, height) / 2`. It can be
assumed that a horizontal scroll bar always has width greater than
height, and vertical scroll bar always has height greater than width.
Even if this is not the case (such as, when it is tiny), again the
behavior would not be different.

This came from the Qt's own implementation for its "Basic" style as is.
Ideally this should be sent to the upstream, but it is not important.