qml: divide radius by two in `ScanProgressBar.qml`
This does not effectively change anything, as the `Rectangle` clamp it
anyway, but it still worth to have.

As the documentation states in `RoundButton`, radius can not be bigger
than half of minimum of width and height:

> To create a completely circular button (the default), use a value that
> is equal to half of the width or height of the button, and make the
> button's width and height identical.

Not dividing by two can be considered a mistake. I specifically did not
use `Math.min(width, height) / 2`, as it can be assumed that this
horizontal progress bar always going to have width bigger than height.
Even if this is not the case, again the behavior would not be different.