qml: divide radius by two in `MusicArtistDelegate.qml`
This does not effectively change the rounding, as the `ImageExt` clamp
it anyway, but it is necessary to have because shadows depend on it.

As the documentation states in `RoundButton`, radius can not be bigger
than half of minimum of width and height:

> To create a completely circular button (the default), use a value that
> is equal to half of the width or height of the button, and make the
> button's width and height identical.

Not dividing by two can be considered a mistake. I specifically did not
use `Math.min(width, height) / 2`, as this image is supposed to have the
same width and height (it is supposed to be a circle).