qt: do not divide radius by two in `RoundedRectangleShadow.frag`
I realized that this is a mistake, the shadow implementation
already operates the same way as `Rectangle` or `ImageExt` does,
so we should not need to divide by two.

I initially thought that we needed this because of unmatched
radii in `MusicAllArtists` between the shadow and the image
when I was experimenting, but turns out it was a bug in the
radius itself which is now fixed by 639a4819.