qml: do not manually set the radius of shadow in `PlayButton.qml`
Instead of dividing radius by two, I just moved the
shadow for it to be a child of the outer circle, so
it can get the correct radius from its parent directly.

Since the shadow uses `z: -1` inherently, we should
not have a stacking issue here. The shadow can be
expected to always rendered under the outer (and
inner) circle.