Revert "qt: do not adjust size in `VideoWindowHandler` with Wayland"
This reverts commit dae15e996aee610d97e35f836f1d6b0f98c1b3d3.

After checking this again, I realized that we need to revert it as
it does the opposite of what we want. Even if DPR is already considered
in `CompositorWayland`, without this revert, the final size end up
being the scaled one (where we want to use the natural size of the
video). It seems that there was something wrong with how I tested it
back then.