qt: compensate for non-video area with auto resize when resizing the interface window
When controls are not overlayed (pinned control bar), the embedded video window has
a non-matching, less size than the interface window. With auto resize, although it
is advertised as "interface resizes to video size" and it is working as expected,
the expectation is actually the embedded video area resizing and the control bar
taking additional space. This practically means that when the interface window is
resized, it needs to compensate for the non-video area.

This should make the behavior consistent with VLC 3.0.