qt: set root item's parent in `CompositorX11UISurface::setContent()`
We are already setting the visual parent, we might as well set the
object parent. It should be acceptable to take ownership here,
unless I'm missing something. Besides, QQuickView also does this.

I did not remove the manual deletion in the destructor, as I was
worried about the ordering there. With the weak pointer, we should
not have double free here (if somehow the window content item is
destroyed before hence it destroyed its object child root item).

This makes it useful to make use of `QObject` utilities such as
`findChild()`.