qt: enforce synchronization in video surface when vout becomes available
This is because even when there is video surface provider bound, without
vout the forwarded requests are not respected in the provider.

Granted, one can argue that it should be the provider's responsibility to
store to forwarded properties from the surface item, and apply them as
soon as vout is available. However, that would be really debatable, because
the availability of vout is public in the provider.