qt: schedule update after setting `ItemHasContents` in video surface
This should not be necessary at all, because when the item starts
having content, Qt should automatically schedule an update anyway
as without the initial update, the item can not paint its content.

However, we don't lose anything by this, so it should be okay as
a precautionary measure in case Qt does not do that.