qt: expose texture view state as properties in `TextureProviderItem`
This can be useful in certain situations, like tiling the source
texture without a layer (thanks to wrap mode).

It should be noted that these are the view states, we are not
modifying the source texture state here.

`ShaderEffectSource` already exposes these for the layer texture
it manages. Though, it seems more limited there, for example it
does not seem to allow disabling filtering completely, or allow
adjusting anisotropy level (which does not mean anything at the
moment but nevertheless).