demux: ps: define what flavor of PS streams can be expected
Between VOB and AOB files the 0xa0->0xaf substreams may be used
with different codec or different encapsulation.
There is nothing in the data to really distinguish them. But when reading
either of those files we know if the file we opened is a dvd, dvda, .VOB,
.AOB or something else.

This will break detection of MLP in AOB files that are not named *.AOB.
But it was already broken in a different way by checking a certain value.

In any case DVD sources will be considered as DVD unless it's marked as
coming from DVD-Audio.

The DVD-Audio support in dvdread is commented out and will have to be
uncommented when DVD-Audio support is added.

(cherry picked from commit 1ae5b0b8d4c67a74d9f98d84aad78b4f38a4cb8c)
- 3.0 uses a block_t rather than buffer+size
- the code around is slightly different
- the 0xa001||0xbda1 check is done on one line ine 3.0