qt: use mipmap filtering by default if source is mipmapped in `TextureProviderItem`
This fixes the regression in 0e5200bc where it made mip map filtering off by default
for mipmapped source textures (unless `mipmap: TextureProviderItem.Linear` is used,
which we do not want in each case explicitly).

This is not really a bug, but I consider this a regression because it is hard to
justify not using mipmap filtering when the texture has mip maps.