qml: do not disable sub-pixel alignment with Qt < 6.2.5 in flickables
This is for the Qt bug QTBUG-103996. Normally we should not care about
this because Qt 6.2.12 is already released open-source, but Qt does
not seem to offer these LTS versions pre-built and Ubuntu 20.04 LTS
still seems to offer Qt 6.2.4, so it can be useful to have this check
for now.