extras: snap: force using local ninja
The vlc-ubuntu-bionic:20250904134924 contains ninja-1.13.0.git.kitware.jobserver-pipe-1.
It supports jobserver as a client like the old kitware fork. But it only
does so when combined with GNU make 4.4.

It seems to mess the job sharing in the CI so use the one that is validated
with old GNU make versions.