avcodec: encoder: remove AV_INPUT_BUFFER_MIN_SIZE usage
It's not supported in FFMpeg 8.0 since 60.40.100.

It used to be used with preallocated packet buffers with
the old encode API, but said API is no more and therefore
there is no reason for this to be public any more.
So deprecate it and use an internal replacement
for the encoders using it as an upper bound for the
size of their headers.

(cherry picked from commit dae7faa0757ee094681c5e7100baca74c007c67c)