qml: do not enable blending if source has no alpha channel in `DualKawaseBlur`
This is useful if the source is a texture generated from a JPEG image, for
example, where there can't be an alpha channel.

PNG images would probably have an alpha channel, whether or not the image
actually uses transparency, so this may not mean much in that case but it is
still good to have.

This should be safe because we are not supporting having transparent border
(like `FastBlur`), so blending should solely depend on the source texture.