qt: add property `isValid` to `TextureProviderObserver`
This is to see if the source texture provider actually
provides a texture, and especially useful if it is
important to know when the texture would be available.

One particular use case for this is with static effects,
such that it makes it possible to apply the effect once
the source provider starts providing the texture.

`Image` as texture provider has `Image.Ready` status,
but it seems that it is not really the best reflection
for whether the texture is ready or not. My understanding
is that it is rather for when the image is loaded, and
not necessarily the texture.