qml: use `MainCtx::intfDevicePixelRatioChanged()` for dpr changes in `PartialEffect`
`TextureProviderItem` does not emit `dprChanged()` signal because Qt apparently
does not forward that to items that do not have `ItemHasContents` flag set.

This fixes the frosted glass effect looking broken at start on Wayland with
fractional scale behind the scan progress bar (when it is immediately shown
upon start) if the progress bar is shown before the fractional scale is
retrieved (where the ceiled value is used in that case, but never updated).