qml: use more reasonable source size for artist banner background image
As long as the interface window is a fraction of the screen width,
the displayed image is smaller than the source size, making it hard
to justify having such big source size. And when the interface window
is big, the details are gone due to blurring, which again makes it
hard to justify having big source (hence image, hence texture) size.

Granted, this is going to make the background more blurry when the
interface window width is bigger than half of the screen size. But,
the increase in blurriness in that case is substantially less than
what the blur effect already achieves. Though with `1/4` it was
rather obvious, so `1/2` should be safer.

This change reduces video and system memory consumption considerably.