qml: make sure the layer is gone when it is not used in `ArtistTopBanner`
This is for the switching case, once the layer is created it would remain,
so setting the `sourceItem` to null while `live` is enabled should help
with this.

Instead of creating the layer once and keeping it around, I believe it
makes more sense to destroy it when it is not used, and re-create when
necessary.