contrib: qt: Apply a patch to fix building QtDeclarative with latest libc++
This fixes https://bugreports.qt.io/browse/QTBUG-140181, fixing
compilation with the latest libc++ (22.x).

This issue is currently only exposed while we're at Qt 6.8; in
Qt 6.9, the code in QtDeclarative that breaks with latest libc++
has been rewritten.

This was the initial fix suggestion in Qt, from
https://codereview.qt-project.org/c/qt/qtbase/+/676607. However,
there were potential binary compatibility concerns with this patch,
so in the end, they ended up fixing it differently (with a larger
number of larger patches - see the bug report above for links to all
those patches). However for the purposes of VLC contribs (statically
linked Qt into the VLC plugins) this simple patch should be fine.

The bug fix in Qt has another part, avoiding potential problematic
code in QtDeclarative for the 6.8 branch, but that fix commit is
not yet public opensource (as support for older Qt branches is
restricted to paying customers, initially). But applying the
fix on qtbase seems to be enough to fix compilation with latest
libc++.