contrib: use curl --retry to download tarballs
If some URLs are flaky we have a better chance of not failing a build.
(like GNU mirrors and git)

We don't need a temporary file, we let curl manage the output file and
overwrite it when necessary. If anything the hash will not match.

(cherry picked from commit b57f13f97e0c395791aeff997a85380e4769d0df)