qml: make use of built-in postprocessing in `FrostedGlassEffect`
This also reverts the frosted glass look back to the time when
we used an additional layer to have exclusion and tint (988b7bb2).
Noise-wise, it should not have any changes.

Thanks to the custom blur effect, we can now have shader-level
adjustments as we prefer directly in the blur effect without
needing an additional layer, similar to Qt's `MultiEffect`.

This also eases the work of the scene graph (batch) renderer.