qt: force stylesheet to be re-evaluated when palette changes
As far as I understand, the colors used by the widgets are evaluated when we
set a stylesheet on an item (even if the style sheet doesn't specify any
colors), if the applicationpalette changes afterward, the colors of a widget
with a custom style is not updated properly. Forcing the stylesheet to be
re-evaluated when the palette changes, forces the colors to be updated.

since Qt 5.7, we can use Qt::AA_UseStyleSheetPropagationInWidgetStyles which
propagates the palette automatically

fix: #29315