qml: use sub-texture for source instead of filtering when feasible in `PartialEffect`
Providing `sourceVisualRect` now makes the source visual to use sub-texture
instead of filtering out the effect part, which is done when blending is
enabled.

When blending is off, since the opaque effect is drawn first, the fragments
of the source visual that are obscured by the effect area should already be
discarded due to early fragment test, provided that the scene graph uses
depth buffer (which is the default). If not, `sourceVisualRect` can be used
to have the same benefit.