freetype: darwin: avoid warning about VLA folded into array
A const size_t is not quite the same as an integer constant and recently
compilers started to warn about this. We can easily avoid this by using
a define, however here we can just use the size of the array itself
and do not need to hardcode any size at all.

Avoids the following compiler warning with recent Clang:
  warning: variable length array folded to constant array as an extension