qt: do not allow attaching multiple video surfaces to the provider
Having multiple video surfaces is technically not possible,
but during scene transitions there can be multiple coexisting
video surface items in the scene.

For example, when switching to the player page, a new video
surface item is created before the old one (pip player) is
destroyed.

This patch makes the video surface item cooperative, that is,
it probes the existing video surface and detaches it first
before attaching itself to the provider. The last one created
becomes the active video surface.