cvdbsub: don't use metadata offset if bigger than whole SPU
ParseMetaInfo() will not parse anything in that case as p == p_end.
The image will use the whole of the SPU buffer, potentially containing
metadata at the end. But the reading in RenderImage() is constrained by the width/height
so it won't use these data.

Fixes https://code.videolan.org/videolan/vlc/-/issues/29325