qt: make sure playlist model recovery is not initialized during shutdown
We are queuing the initialization to not slow down the start. However,
if -for any reason- the application is already closing before the queued
slot was executed, then we should make sure to not initialize the model
recovery as not only it would not make any sense, but also may cause
null pointer dereference because the model or settings may be gone.