qml: do not use `positionViewAtIndex()` on key navigation unless necessary in `ListViewExt`
Normally the view itself does positioning, provided that `highlightFollowsCurrentItem` is
enabled (default). When `positionViewAtIndex()` is used, that is not the case, but as the
note tells, the item may end up behind the header or footer without it. This is not a Qt
bug, because the option is named as overlay, which means the items are supposed to go
behind.

Ideally the `positionViewAtIndex()` should not be used at all, because the correct way of
having header or footer that are neither meant to be overlay or inline seems to be having
them outside the view, such as how it is done in playqueue. Unfortunately Qt does not seem
to offer this option built-in in the views.