qt: inline the CONNECT macro into the remaining callers
Some call sites cannot use type-safe connections, due to either using a pointer
to an incomplete type as slot argument, or dynamic typing. However these are
the exception now. For that reason string-based connection is performed
explicitly when required. This is part of a spiritual backport of [1], but has
been written from scratch due to excessive conflicts.

[1] https://code.videolan.org/videolan/vlc/-/commit/eb0b785fdf5396b429fb039b0c92d44f768fe65c