tools: keep the config.sub/config.guess locally
The GNU and github URLs are not reliable. And since we always download
the same version we might copy it locally.

We keep the source URL around for when we need to update the files.

They are installed in <tools_build>/extras/tools/autoconf-vlc and picked
from there by contrib and bootstrap.