vlm: remove "schedule"
The schedule feature was essentially just for fun at the ECP campus and
never really worked. In practice people do (or at least did) schedule
broadcasts with VLC, but they that with soem kind of higher-level
management process starting and stopping VLC (and/or sending VLM broadcast
commands) - not with VLM schedules.

Note that this does _not_ remove VLM broadcasts. That code is still
required to support the streaming output bridge and mosaic bridge or simply
to test multiple inputs in a single process. Whilst #27733 implies that
VLM broadcasts are not currently working either, they are outside the
scope of this changeset either way.

Known problems with the schedules:
- Neither daylight saving and time zones are not supported so this fails
  at least twice a year if you have daylight savings.
- (Day of) Month-based scheduling actually worked (the calculation was
  just plain wrong).
- Clock adjustments broke it.
- It required realtime clock-based condition variable sleep which most
  operating systems don't support.

Closes #25063, #25064, #25065, #25066, #25067, #25068, #25069, #25070.