contrib: gnutls: remove unused GNUTLS_ENV
MAKECONFIGURE is equivalent to
> cd $(BUILD_DIR) && $(HOSTVARS) $(BUILD_SRC)/configure $(HOSTCONF)

The GNUTLS_ENV is only applied on the cd command and useless.
Since it's building fine on Android we don't need it.

This is the only call with an environment variable before MAKECONFIGURE.

The hack was originally added in bead358b379601eb4be6d9914bd3379ff73b50ba
and made useless in cfeea778135bf125ad2720a97199b7f1233b07af.