.tx: add configuration to update po files
This is a configuration file for the tx command line tool to
automatically pull updated translations from the transifex platform to
the correct places.

After the pull, it is necessary to correct the headers manually and to
perform "msgcat --width=79 po/*.po -o -" to achieve the intended
formatting.

Perform the pull with "tx pull --use-git-timestamps --force"