Revert "demux:mkv: fix wrong value reset after clean of an array"
This reverts commit d1d41b3f849d4e260c1ec09c5b1bb53db7f8107f.

The i_subpackets value represents the size of the p_subpackets array.
It should not change during the lifetime of that array.

The i_subpacket value is the count of actually filled elements in the array.
When the array is fully loaded we can send the blocks for decoding.

The fix for the hacker one file should be different.

(cherry picked from commit f9500c236765ab64cb00cc4fcbd58db4ec02a193)