qt: calculate atlas coordinate manually in `ImageExt`
Instead of relying on `supportsAtlasTextures`, we now
calculate the atlas coordinate manually because the
proper coordinate is provided after the texture is
already painted during initialization otherwise.

We can do this without worrying about atlas textures
getting detached from the atlas because having
`qt_SubRect_source` in the uniform block serves the
same purpose, except the coordinate in the atlas must
be calculated manually in that case.