qml: display the visual when its source texture is ready in `DualKawaseBlur`
Akin to 75c368c2, if the source texture is not available, a black/transparent
texture is generated and used instead. Not only this is a waste, but also it
can behave like `ViewBlockingRectangle` if blending is disabled, which can
cause glitches.

We can simply make the painter delegate visible when its source is ready,
thanks to aa2b1ec6.