qt: disable pip player and backdrop blur if there is no depth buffer with rhi
"View blocking", or not letting a certain part of the scene being
painted depends on a graphics optimization that is only possible
with depth buffer.

We could theoretically improve `ViewBlockingRectangle` by using a
custom render node and clear what was already rendered beneath when
there is no depth buffer, similar to how it is already done in
software mode, but that would be bad for maintenance because of
needing to support multiple graphics backends. This approach would
be more robust, but so far relying on depth buffer has not really
been problematic.