qt: disable rcc zstd rather than force the flag when qt runtime doesn't support it
The flag was added in 144757e5f13753fec538893505bb03aaae72b276 because qt
runtime may be compiled without zstd support while rcc may support it.

But the other scenario exists with rcc with no zstd support and qt with zstd
support.

Instead of forcing zstd we can disable it in rcc when Qt doesn't support it

* zstd is enabled by default when rcc support it, so forcing it is redundant if
  both rcc and Qt supports it
* --no-zstd flags exists even if rcc doesn't support zstd

this is the strategy used in Qt `mkspecs/features/resources.prf`