Revert "qt: calculate atlas coordinate manually in `ImageExt`"
This reverts commit e824b12ab741a1791929ac3a541198ea19ffe25b.

Unfortunately even though `supportsAtlasTextures` is not
required to support atlas textures (meaning, operating on
the atlas texture without needing to call `QSGTexture::
removedFromAtlas()`), not having it set breaks batch
rendering.

The reasoning for this is not really clear, because Qt
could still batch `ShaderEffect`s that support atlas
textures through the uniform sub-rect define, provided
that there is only one source texture and it is the
same (the atlas texture). The only difference of
`supportsAtlasTextures` in this case the texture coordinate
provided to the shader being atlas-relative, which
should not have an effect on batch rendering.

That being said, shadows used underneath, possibly due
to overlapping between other items, already seem to
prevent batching of the delegate image. However, that
is a different concern and should be handled separately.

This brings back the initialization glitch, but batch
rendering is considered more important than the glitch,
and the glitch is not so bad that warrants a workaround
at the moment.