decoder: fix vctx check
If vctx == NULL (no hw-dec) and created_vout == 0 (not format change),
do not request a new vout (via input_resource_RequestVout()).

This was not causing any restart as input_resource_RequestVout() would
re-use the same one, but it allows to do less check and less lock/unlock
for each frames that is pushed.