qml: calculate painted size manually at all times in `ImageExt`
Calculating painted size turned out to be trivial, so we can
calculate the size ourselves at all times and not rely on
`Image`'s calculation because although it provides the painted
size when it is hidden at the moment, it is not guaranteed to
be the case in the future (as noted in the comments previously).

This allows us to remove the warning regarding supporting
fill modes that are different than preserve aspect crop (as
for that we already do the necessary calculation ourselves in
`cropRate`).